Specify files in package.json
Whitelist only the `lib` directory for npm publication. All other files 
will not be published as part of the npm package (with the exception of 
the entry JS file, README, LICENSE, and CHANGELOG).

This will, for example, avoid publishing our future tests directory.

This was inspired by the following article: 
https://medium.com/@jdxcode/for-the-love-of-god-dont-use-npmignore-f93c08909d8d