Cypress: Build and deploy as part of test
In order to test different project setups (e.g. without 404, with
static 404, with dynamic 404, ...), we need to build and deploy the
project as part of each Cypress spec file.

The building and deployment happens in a spec file's before hook. We
track the PID of the server that was started (if deploying locally) and
shut it down on completion.