_redirects file: Remove path param from  Netlify functions
In the past, there was a difference between how Netlify and netlify dev
handled redirects. Only on the Netlify platform would the original path
be available to the function (e.g. /mypage). Locally, only the function
path would be available (e.g. /.netlify/functions/nextRouter).

This caused a problem when we want to identify the page to render in
nextRouter. Locally, we did not know the original path as requested by
the user. To work around that, we passed the page's path as a parameter
to the function call.

This behavior of netlify dev was fixed in
https://github.com/netlify/cli/pull/713