Add support for SSR pages using getServerSideProps
For pages that use getServerSideProps, we create a Netlify Function
just like we do for pages with getInitialProps. In addition, we also
add an extra redirect that allows getting the page props as JSON.
This is when navigating to the page on the client-side.

See: https://github.com/FinnWoelm/next-on-netlify/issues/7