Simplify build output for SSG pages
Combine the line for page HTML and page JSON into a single line per page
and move output for SSG pages with fallback to section for SSR pages.

This way, we can eliminate the nesting from SSG pages (we no longer have
subsection 1, 2, and 3). This will allow us to more easily implement the
maxLogLines option for limiting output for each section to a specific
number of lines.